Split routes/analysis.ts by resource; mount all under projectResolver
analysis.ts had grown into a ~700-line catch-all. The original split
(analyze.ts vs analysis.ts) was structured around a DB-readiness
distinction: analyze bootstrapped the project's DB, analysis assumed
it was up. That distinction evaporated with the file-based store —
`.truecourse/` either exists on disk or it doesn't, no async init.
POST /diff-check already crossed the line (a long-running mutation
living in the "analysis" router), confirming the original seam was
stale.

Resource-based split instead:

  analyze.ts    — long-running mutations sharing the 202+sockets
                  contract: POST /analyze, POST /analyze/cancel,
                  POST /diff-check, GET /diff-check
  analyses.ts   — stored-analysis CRUD: GET /analyses,
                  GET /analyses/:id/usage, DELETE /analyses/:id
                  (carries the rebuild-LATEST helper)
  graph.ts      — graph rendering + UI state: GET /graph,
                  PUT/DELETE /graph/positions, PUT /graph/collapsed
  files.ts      — filesystem + git surface: GET /files,
                  GET /file-content, GET /changes

All four now mount under `projectResolver` uniformly — previously
analyze.ts skipped it as part of the bootstrap exception.

`findAnalysisFilename` (used by analyses + graph) moved to
`lib/analysis-store.ts` so both routers can share it without
duplication.

No behavior changes — same endpoints, same responses, same shape.
Tests stay green (3375 passed).
